# docs/openAPI_model.yml
openapi: 3.0.3
info:
  title: HR Event & Leave Management API
  version: 1.0.0
  description: Backend API for managing employees, HR events, leave requests, and reporting.
servers:
  - url: http://localhost:8000/api
paths:
  /employees:
    get:
      summary: List employees
      responses:
        "200":
          description: List of employees
    post:
      summary: Create new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
      responses:
        "201":
          description: Employee created
  /employees/{id}:
    get:
      summary: Get employee by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee details
    put:
      summary: Update employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
      responses:
        "200":
          description: Employee updated
    delete:
      summary: Delete employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Employee deleted
...
components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
        seniority:
          type: string
...